Add support for checking out a specific git revision by ID
Add the option 'ipbb add git ... -r XXX' to check out revision XXX
of the specified repo. Both long and short revision hashes are
supported. Since the existence of the hash cannot be verified before-
hand (at least not without first cloning the repo) it is simply
assumed to exist. In case the checkout goes wrong, a message is
issued and the repo is reverted to the master branch.