Improve support for .xdc and .tcl constraints files
So far, ipbb treated xdc and tcl constraints files differently. XDC
files would follow the default Vivado behaviour, and be used for
both synthesis and implementation. TCL files would only be used in
implementation.

These changes make both xdc and tcl constraints files follow the
default Vivado behaviour, and add an optional '--usefor' flag in
the dependency files to specify where/when to use the file.
For example: '--usefor implementation', to only use the file in the
implementation step, or '--usefor synthesis,implementation', to
use the file in both the synthesis and the implementation step.