Apply the work-around for the 'IPBB terminal' hanging to all OS's
This problem originally seemed limited to RedHat 8-based systems.
Now, however, it also shows up when using CC7-based setups. For
example when using the Vivado 2022.2 image from
https://gitlab.cern.ch/cce/docker_build/
(albeit on an Alma 8.9 host).

So, especially since the work-around is relatively efficient, it
is easier to simply always apply it instead of trying to find a
heuristic to detect which systems are sensitive to this problem.