Avoid 'cropping' of contents of wide dependency tables
The current behaviour is to 'crop' strings in rows that would exceed
the terminal width and append ellipsis. This looks nice, and may be
very grep-friendly, but it removes some interesting information. As
in: it tends to remove the tail ends of long path/file names.
This change makes these occurrences 'fold over' to the next line
instead. Not as pretty, but it retains all information.