Make sure that 'ipbb vivado package' always produces the package in the same location
At the moment, the 'vivado package' command generates the package in the current
directory. So when one executes this command from outside the project directory,
the package directory is created outside the project directory too.

Assuming this is _not_ an explicit design choice, this modification ensures
that the package is always created in the project directory.